Move tests into 'tests' directory
* berg_c64, berg_f64 are now private
