+++ src/berg_c64.rs
@@ -75,66 +75,3 @@ pub fn berg(lag: usize, delta: f64, y: &ArrayView1<Complex64>, q: &ArrayView1<f6
 
     psd
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ndarray::stack;
-    use ndarray::Array1;
-    use ndarray::Axis;
-    use ndarray_npy::write_npy;
-
-    use num_complex::Complex64;
-
-    #[test]
-    fn sgn() {
-        // Prepare dataset
-        let xs = Array1::<f64>::linspace(-5., 4.99, 1000);
-        let mut ys = Array1::<Complex64>::zeros(1000);
-        for (x, y) in xs.iter().zip(ys.iter_mut()) {
-            if *x > 0. {
-                *y = Complex64::new(1., 0.);
-            } else {
-                *y = Complex64::new(-1., 0.);
-            }
-        }
-
-        // Compute PSD
-        let q = Array1::<f64>::logspace(10., -5., 2., 100);
-        let f = berg(100, xs[1] - xs[0], &ys.view(), &q.view());
-
-        // Save result
-        assert!(write_npy("sgn_c.npy", &stack![Axis(1), xs, ys.map(|x| x.re)]).is_ok());
-        assert!(write_npy("sgn_result_c.npy", &stack![Axis(1), q, f]).is_ok());
-    }
-
-    #[test]
-    fn gauss() {
-        // Prepare dataset
-        let xs = Array1::<f64>::linspace(-10., 10., 1000);
-        let ys = xs.map(|x| Complex64::new((-1.0 * x.powf(2.)).exp(), 0.));
-
-        // Compute PSD
-        let q = Array1::<f64>::linspace(0., 3., 100);
-        let f = berg(361, xs[1] - xs[0], &ys.view(), &q.view());
-
-        // Save result
-        assert!(write_npy("gauss_c.npy", &stack![Axis(1), xs, ys.map(|x| x.re)]).is_ok());
-        assert!(write_npy("gauss_result_c.npy", &stack![Axis(1), q, f]).is_ok());
-    }
-
-    #[test]
-    fn exp() {
-        // Prepare dataset
-        let xs = Array1::<f64>::linspace(-20., 20., 2000);
-        let ys = xs.map(|x| Complex64::new((-1.0 * x.abs()).exp(), 0.));
-
-        // Compute PSD
-        let q = Array1::<f64>::linspace(-1., 1.5, 100);
-        let f = berg(100, xs[1] - xs[0], &ys.view(), &q.view());
-
-        // Save result
-        assert!(write_npy("exp_c.npy", &stack![Axis(1), xs, ys.map(|x| x.re)]).is_ok());
-        assert!(write_npy("exp_result_c.npy", &stack![Axis(1), q, f]).is_ok());
-    }
-}
