+++ src/berg_f64.rs
@@ -70,82 +70,3 @@ pub fn berg(lag: usize, delta: f64, y: &ArrayView1<f64>, q: &ArrayView1<f64>) ->
 
     psd
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ndarray::arr1;
-    use ndarray::s;
-    use ndarray::stack;
-    use ndarray::Array1;
-    use ndarray::Axis;
-    use ndarray_npy::write_npy;
-
-    use num_complex::Complex64;
-
-    #[test]
-    fn complex_array_test() {
-        let f = Array1::<Complex64>::ones(1);
-        assert_eq!(f[0], Complex64::new(1., 0.));
-    }
-
-    #[test]
-    fn slice_reverse_test() {
-        let a = Array1::<f64>::linspace(0., 9., 10);
-        let b = arr1(&[9., 8., 7.]);
-        let c = a.slice(s![7..10;-1]);
-        assert_eq!(c, b);
-    }
-
-    #[test]
-    fn sgn() {
-        // Prepare dataset
-        let xs = Array1::<f64>::linspace(-5., 4.99, 1000);
-        let mut ys = Array1::<f64>::zeros(1000);
-        for (x, y) in xs.iter().zip(ys.iter_mut()) {
-            if *x > 0. {
-                *y = 1.;
-            } else {
-                *y = -1.;
-            }
-        }
-
-        // Compute PSD
-        let q = Array1::<f64>::logspace(10., -5., 2., 100);
-        let f = berg(100, xs[1] - xs[0], &ys.view(), &q.view());
-
-        // Save result
-        assert!(write_npy("sgn.npy", &stack![Axis(1), xs, ys]).is_ok());
-        assert!(write_npy("sgn_result.npy", &stack![Axis(1), q, f]).is_ok());
-    }
-
-    #[test]
-    fn gauss() {
-        // Prepare dataset
-        let xs = Array1::<f64>::linspace(-10., 10., 1000);
-        let ys = xs.map(|x| (-1.0 * x.powf(2.)).exp());
-
-        // Compute PSD
-        let q = Array1::<f64>::linspace(0., 3., 100);
-        let f = berg(361, xs[1] - xs[0], &ys.view(), &q.view());
-
-        // Save result
-        assert!(write_npy("gauss.npy", &stack![Axis(1), xs, ys]).is_ok());
-        assert!(write_npy("gauss_result.npy", &stack![Axis(1), q, f]).is_ok());
-    }
-
-    #[test]
-    fn exp() {
-        // Prepare dataset
-        let xs = Array1::<f64>::linspace(-20., 20., 2000);
-        let ys = xs.map(|x| (-1.0 * x.abs()).exp());
-
-        // Compute PSD
-        let q = Array1::<f64>::linspace(-1., 1.5, 100);
-        let f = berg(100, xs[1] - xs[0], &ys.view(), &q.view());
-
-        // Save result
-        assert!(write_npy("exp.npy", &stack![Axis(1), xs, ys]).is_ok());
-        assert!(write_npy("exp_result.npy", &stack![Axis(1), q, f]).is_ok());
-    }
-}
